Moved convolution code to a separate file
